--- app/src/main/java/com/example/juegita/ProfileScreen.kt
@@ -1,221 +1,0 @@
-package com.example.juegita
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.AccountCircle
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material.icons.filled.Edit
-import androidx.compose.material.icons.filled.Email
-import androidx.compose.material.icons.filled.Lock
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.NavController
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun UserProfileScreen(navController: NavController) {
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = {
-                    Box(
-                        modifier = Modifier.fillMaxWidth(),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        val gradientColors = listOf(Color.Cyan, Color.Magenta)
-                        Text(
-                            text = "Perfil de usuario",
-                            fontSize = 26.sp,
-                            fontWeight = FontWeight.Bold,
-                            modifier = Modifier.align(Alignment.Center),
-                            style = TextStyle(
-                                brush = Brush.linearGradient(colors = gradientColors)
-                            )
-                        )
-                    }
-                },
-                navigationIcon = {
-                    IconButton(onClick = { navController.navigate("settings") }) {
-                        Icon(
-                            imageVector = Icons.Default.ArrowBack,
-                            contentDescription = "Ir a la configuración",
-                            tint = Color.Black,
-                            modifier = Modifier
-                                .size(40.dp)
-                                .padding(end = 1.dp)
-                        )
-                    }
-                },
-                colors = TopAppBarDefaults.topAppBarColors(
-                    containerColor = Color(0xFF6200EA)
-                )
-            )
-        },
-        content = { innerPadding ->
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(Color.White)
-                    .padding(innerPadding),
-                contentAlignment = Alignment.TopCenter
-            ) {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(16.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    // Icono
-                    Icon(
-                        imageVector = Icons.Default.AccountCircle,
-                        contentDescription = "User Icon",
-                        tint = Color.Black,
-                        modifier = Modifier.size(194.dp)
-                    )
-
-                    // Nombre de usuario
-                    Text(
-                        text = "Username",
-                        fontSize = 24.sp,
-                        fontWeight = FontWeight.Bold,
-                        color = Color.Black
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-
-                    // Nombre
-                    Text(
-                        text = "Nombre",
-                        fontSize = 16.sp,
-                        color = Color.Black,
-                        textAlign = TextAlign.Center,
-                        modifier = Modifier.padding(horizontal = 16.dp)
-                    )
-                    Spacer(modifier = Modifier.height(16.dp))
-
-                    // Sección de editar perfil
-                    ExtendedFloatingActionButton(onClick = { navController.navigate("editar-perfil") },
-                        modifier = Modifier.fillMaxWidth(),
-                        containerColor = Color.White,
-                        contentColor = Color.Black,
-                    ) {
-                        Icon(Icons.Filled.Edit, "")
-                        Text(
-                            text =" EDITAR PERFIL",
-                            fontSize = 22.sp,
-                            fontWeight = FontWeight.Bold
-                        )
-                    }
-                    Spacer(modifier = Modifier.height(10.dp))
-                    // Sección de editar perfil
-                    ExtendedFloatingActionButton(onClick = { /* Acción al hacer clic en el botón*/ },
-                        modifier = Modifier.fillMaxWidth(),
-                        containerColor = Color.White,
-                        contentColor = Color.Black,
-                    ) {
-                        Icon(Icons.Filled.Lock, "")
-                        Text(
-                            text =" CAMBIAR CONTRASEÑA",
-                            fontSize = 22.sp,
-                            fontWeight = FontWeight.Bold
-                        )
-                    }
-                    Spacer(modifier = Modifier.height(10.dp))
-                    // Sección de actividades recientes
-                    AssistChip(
-                        onClick = {},
-                        label = {
-                            Box(
-                                modifier = Modifier.fillMaxWidth(),
-                                contentAlignment = Alignment.Center
-                            ) {
-                                Text(
-                                    text = "Actividad reciente",
-                                    color = Color.Black,
-                                    fontSize = 22.sp,
-                                    fontWeight = FontWeight.Bold,
-                                    modifier = Modifier.align(Alignment.Center),
-                                )
-                            }},
-                        modifier = Modifier.fillMaxWidth()
-                            .height(45.dp),
-                        elevation = AssistChipDefaults.assistChipElevation(1.dp),
-                        border = null,
-                        colors = AssistChipDefaults.assistChipColors(Color.White)
-
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-                    RecentActivitiesSection()
-
-                    Spacer(modifier = Modifier.height(16.dp))
-
-                    Spacer(modifier = Modifier.height(16.dp))
-
-                    // Enlaces a redes sociales
-                    Text(
-                        text = "Amigos",
-                        fontSize = 20.sp,
-                        fontWeight = FontWeight.Bold,
-                        color = Color.Black,
-                        modifier = Modifier.align(Alignment.Start)
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-
-                }
-            }
-        }
-    )
-}
-
-@Composable
-fun RecentActivitiesSection() {
-    Column(modifier = Modifier.fillMaxWidth()) {
-        ActivityCard(activity = "Jugó Memorama - 10/10/2024")
-        ActivityCard(activity = "Completó el curso de Kotlin - 09/10/2024")
-        ActivityCard(activity = "Publicó una nueva foto - 08/10/2024")
-    }
-}
-
-@Composable
-fun ActivityCard(activity: String) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(4.dp),
-        shape = RoundedCornerShape(8.dp),
-        elevation = CardDefaults.cardElevation(4.dp),
-        colors = CardDefaults.cardColors(Color.White)
-    ) {
-        Text(
-            text = activity,
-            modifier = Modifier.padding(16.dp),
-            fontSize = 16.sp,
-            color = Color.Gray
-        )
-    }
-}
-
-
-@Preview(showBackground = true)
-@Composable
-fun PreviewUserProfileScreen() {
-    UserProfileScreen(navController = NavController(LocalContext.current))
-}
