--- app/src/main/java/com/example/juegita/RegisterScreen.kt
@@ -1,310 +1,0 @@
-package com.example.juegita
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.AccountCircle
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.AddCircle
-import androidx.compose.material.icons.filled.Create
-import androidx.compose.material.icons.filled.Email
-import androidx.compose.material.icons.filled.Info
-import androidx.compose.material.icons.filled.Lock
-import androidx.compose.material.icons.filled.Person
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.NavController
-import androidx.navigation.compose.rememberNavController
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun RegisterScreen(navController: NavController, modifier: Modifier = Modifier ) {
-    var email by remember { mutableStateOf("") }
-    var username by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-    var confirmPassword by remember { mutableStateOf("") }
-    var passwordVisible by remember { mutableStateOf(false) } // Controla la visibilidad de la contraseña
-    var passwordsMatch by remember { mutableStateOf(true) }
-    var checkedRemember by remember { mutableStateOf(false) }
-    // Background Color
-    Box(
-        modifier = modifier
-            .fillMaxSize()
-            .background(Color(0xFF6200EA)) // Light Blue background
-            .padding(5.dp)
-    ) {
-//        Image(
-//            painter = painterResource(R.drawable.fondo_neon),
-//            contentDescription = "Background Image",
-//            modifier = Modifier
-//                .fillMaxSize(),
-//            contentScale = ContentScale.Crop // Escala la imagen para cubrir todo el espacio
-//        )
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center,
-            modifier = Modifier.fillMaxSize()
-        ) {
-            // Icono grande circular encima de todo
-            Box(
-                contentAlignment = Alignment.Center,
-                modifier = Modifier
-                    .size(140.dp) // Tamaño del ícono
-                    //.clip(CircleShape) // Forma circular
-                    //.background(Color.Black)
-                    .padding(4.dp)
-            ) {
-                Icon(
-                    painter = painterResource(R.drawable.person_add),
-                    contentDescription = "Create Icon",
-                    tint = Color(Color.White.value),
-                    modifier = Modifier.size(194.dp)
-                )
-            }
-            val gradientColors = listOf(Color.Cyan, Color.Magenta)
-
-            Text(
-                text = "Crear nueva cuenta",
-                fontSize = 30.sp,
-                fontWeight = FontWeight.Bold,
-                style = TextStyle(
-                    brush = Brush.linearGradient(colors = gradientColors)
-                )
-            )
-
-
-            Spacer(modifier = Modifier.height(32.dp))
-
-            // Username Input
-            Row (
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .background(Color.White)
-                    .padding(8.dp),
-                verticalAlignment = Alignment.CenterVertically, // Alinear verticalmente al centro
-            ){
-                Icon(
-                    imageVector = Icons.Default.AccountCircle,
-                    contentDescription = "Username",
-                    tint = Color.Black,
-                    modifier = Modifier
-                        .size(40.dp)
-                        .padding(end = 8.dp)
-                )
-                TextField(
-                    value = username,
-                    onValueChange = {username = it},
-                    label = { Text("Username") },
-                    modifier = Modifier.fillMaxWidth(),
-                    colors = TextFieldDefaults.textFieldColors(
-                        containerColor = Color.White,
-                        focusedIndicatorColor = Color.Green,
-                        unfocusedIndicatorColor = Color.LightGray
-                    )
-                )
-            }
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            // Email Input
-            Row (
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .background(Color.White)
-                    .padding(8.dp),
-                verticalAlignment = Alignment.CenterVertically,
-            ){
-                Icon(
-                    imageVector = Icons.Default.Email,
-                    contentDescription = "Correo",
-                    tint = Color.Black,
-                    modifier = Modifier
-                        .size(40.dp)
-                        .padding(end = 8.dp)
-                )
-                TextField(
-                    value = email,
-                    onValueChange = {email = it},
-                    label = { Text("Correo") },
-                    keyboardOptions = KeyboardOptions.Default.copy(keyboardType = KeyboardType.Password),
-                    modifier = Modifier.fillMaxWidth(),
-                    colors = TextFieldDefaults.textFieldColors(
-                        containerColor = Color.White,
-                        focusedIndicatorColor = Color.Green,
-                        unfocusedIndicatorColor = Color.LightGray
-                    )
-                )
-            }
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            // Campo de contraseña
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .background(Color.White)
-                    .padding(8.dp),
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Lock,
-                    contentDescription = "Password",
-                    tint = Color.Black,
-                    modifier = Modifier
-                        .size(40.dp)
-                        .padding(end = 8.dp)
-                )
-                TextField(
-                    value = password,
-                    onValueChange = { password = it },
-                    label = { Text("Contraseña") },
-                    visualTransformation = if (passwordVisible) VisualTransformation.None else PasswordVisualTransformation(),
-                    keyboardOptions = KeyboardOptions.Default.copy(keyboardType = KeyboardType.Password),
-                    modifier = Modifier.fillMaxWidth(),
-                    trailingIcon = {
-                        val image = if (passwordVisible)
-                            Icons.Default.Info
-                        else
-                            Icons.Default.Info
-
-                        IconButton(onClick = { passwordVisible = !passwordVisible }) {
-                            Icon(imageVector = image, contentDescription = null)
-                        }
-                    },
-                    colors = TextFieldDefaults.textFieldColors(
-                        containerColor = Color.White,
-                        focusedIndicatorColor = Color.Green,
-                        unfocusedIndicatorColor = Color.LightGray
-                    )
-                )
-            }
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            // Campo de confirmación de contraseña
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .background(Color.White)
-                    .padding(8.dp),
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Lock,
-                    contentDescription = "Confirmar contraseña",
-                    tint = Color.Black,
-                    modifier = Modifier
-                        .size(40.dp)
-                        .padding(end = 8.dp)
-                )
-                TextField(
-                    value = confirmPassword,
-                    onValueChange = {
-                        confirmPassword = it
-                        passwordsMatch = it == password // Verifica si las contraseñas coinciden
-                    },
-                    label = { Text("Confirmar Contraseña") },
-                    visualTransformation = if (passwordVisible) VisualTransformation.None else PasswordVisualTransformation(),
-                    keyboardOptions = KeyboardOptions.Default.copy(keyboardType = KeyboardType.Password),
-                    modifier = Modifier.fillMaxWidth(),
-                    trailingIcon = {
-                        val image = if (passwordVisible)
-                            Icons.Default.Info
-                        else
-                            Icons.Default.Info
-
-                        IconButton(onClick = { passwordVisible = !passwordVisible }) {
-                            Icon(imageVector = image, contentDescription = null)
-                        }
-                    },
-                    colors = TextFieldDefaults.textFieldColors(
-                        containerColor = Color.White,
-                        focusedIndicatorColor = Color.Green,
-                        unfocusedIndicatorColor = Color.LightGray
-                    )
-                )
-            }
-
-            // Mensaje de error si las contraseñas no coinciden
-            if (!passwordsMatch) {
-                Text(
-                    text = "Las contraseñas no coinciden.",
-                    color = Color.Red,
-                    modifier = Modifier.padding(top = 8.dp)
-                )
-            }
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier
-                    .padding(top = 4.dp)
-                    .fillMaxWidth()
-            ) {
-                Checkbox(
-                    checked = checkedRemember,
-                    onCheckedChange = { checkedRemember = it },
-                    colors = CheckboxDefaults.colors(
-                        checkedColor = Color.Green,
-                        uncheckedColor = Color.White,
-                        checkmarkColor = Color.White
-                    )
-                )
-                TextButton(onClick = { navController.navigate("terminos") }) {
-                    Text(
-                        text = "Estoy de acuerdo con los terminos y condiciones",
-                        color = Color.Yellow,
-                        fontSize = 12.sp
-                    )
-                }
-            }
-
-            Spacer(modifier = Modifier.height(24.dp))
-
-            // Register Button
-            Button(
-                onClick = { navController.navigate("minijuegos") },
-                modifier = Modifier.fillMaxWidth(),
-                colors = ButtonDefaults.buttonColors(Color(0xFFFF5722)) // Orange Button
-            ) {
-                Text(text = "Registrarse", color = Color.White, fontSize = 19.sp)
-            }
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            // Go to Login Button
-            TextButton(onClick = { navController.navigate("inicio-sesion") }) {
-                Text("¿Ya tienes una cuenta? Inicia sesión!", color = Color.Yellow)
-            }
-        }
-    }
-}
-
-@Preview(showBackground = true, showSystemUi = true)
-@Composable
-fun PreviewRegisterScreen() {
-    RegisterScreen(navController = rememberNavController())
-}
-
