--- app/src/main/java/com/example/juegita/ui/theme/ForgotPasswordScreen.kt
@@ -1,2 +1,0 @@
-package com.example.juegita.ui.theme
-
