--- app/src/main/java/com/example/juegita/ui/theme/ProjectInfoScreen.kt
@@ -1,2 +1,0 @@
-package com.example.juegita.ui.theme
-
